Replace old tests system into new generic oriented
