+++ src/lib.rs
@@ -21,7 +21,6 @@ mod raw_place;
 mod utils;
 
 pub(crate) use raw_place::RawPlace;
-use std::mem::MaybeUninit;
 pub use {
     alloc::Alloc,
     file_mapped::FileMapped,
@@ -127,137 +126,4 @@ impl<T> Default for System<T> {
     fn default() -> Self {
         Self::new()
     }
-}
-
-#[test]
-fn miri() {
-    pub fn inner<M: RawMem>(mut mem: M, val: M::Item) -> Result<()>
-    where
-        M::Item: Clone + PartialEq,
-    {
-        const GROW: usize = if cfg!(miri) { 100 } else { 10_000 };
-
-        for _ in 0..10 {
-            mem.grow_filled(GROW, val.clone())?;
-        }
-        assert!(mem.allocated() == vec![val; GROW * 10]);
-
-        for _ in 0..10 {
-            mem.shrink(GROW)?;
-        }
-        assert_eq!(mem.allocated().len(), 0);
-
-        Ok(())
-    }
-
-    let val = "foo".to_string();
-
-    inner(Global::new(), val.clone()).unwrap();
-    inner(System::new(), val.clone()).unwrap();
-
-    #[cfg(not(miri))]
-    inner(TempFile::new().unwrap(), val).unwrap();
-}
-
-#[cfg(test)]
-macro_rules! define_impls {
-    (impl RawMem: {
-        $($ctor:expr /* -- */ $(=> in $cfg:meta)? ),+ $(,)?
-    } for [
-        $($test:path as $name:ident),* $(,)?
-    ]) => {
-        define_impls! { @loop
-            [/* empty result */]
-            [ $($ctor $(=> $cfg)? )*]
-            [ $($test as $name |)* ]
-        }
-    };
-
-   (@loop [ $($result:tt)* ] // result accumulation
-           [ $($ctor:expr $(=> $cfg:meta)? )* ] // each ctor with our cfg `not(miri)`
-           [ $test:path as $name:ident | $($tail:tt)* ] // match test with name + tail
-    ) => {
-        define_impls! { @loop
-            [
-                $($result)*
-
-                #[test]
-                fn $name() {
-                    $( $(#[cfg($cfg)])? $test($ctor);)*
-                }
-            ]
-            [$($ctor $(=> $cfg)? )*]
-            [ $($tail)* ]
-        }
-    };
-
-    (@loop [ $($result:tt)* ] [ $($_:tt)* ] [ /* tests still coming */ ] ) => {
-        $($result)*
-    };
-}
-
-#[cfg(test)]
-define_impls! {
-    impl RawMem: {
-        Global::<u32>::new(),
-        System::<u32>::new(),
-        TempFile::<u32>::new().unwrap() => in not(miri),
-    } for [
-        grow as grow_test,
-        grow_with as grow_with_test,
-        allocated as allocated_test,
-        shrink as shrink_test,
-        grow_filled as grow_filled_test,
-        grow_zeroed as grow_zeroed_test,
-        grow_assumed as grow_assumed_test,
-    ]
-}
-
-fn grow<T>(mut mem: impl RawMem<Item = T>) {
-    unsafe {
-        mem.grow(10, |_uninit| {}).expect("error");
-    }
-    assert!(mem.allocated().len() == 10);
-}
-
-fn grow_with<T: Copy + From<u8>>(mut mem: impl RawMem<Item = T>) {
-    let value = T::from(2);
-    mem.grow_with(10, || value).expect("grow");
-    assert!(mem.allocated().len() == 10);
-}
-
-fn allocated<T: Copy + From<u8>>(mut mem: impl RawMem<Item = T>) {
-    assert_eq!(mem.allocated().len(), 0);
-    let value = T::from(2);
-    mem.grow_with(10, || value).expect("grow");
-    assert!(mem.allocated().len() == 10);
-}
-
-fn shrink<T: Copy + From<u8>>(mut mem: impl RawMem<Item = T>) {
-    let value = T::from(2);
-    mem.grow_with(10, || value).expect("grow");
-    assert!(mem.allocated().len() == 10);
-
-    mem.shrink(5).expect("shrink");
-    assert!(mem.allocated().len() == 5);
-}
-
-fn grow_filled<T: Copy + From<u8>>(mut mem: impl RawMem<Item = T>) {
-    let value = T::from(2);
-    mem.grow_filled(10, value).expect("grow");
-    assert!(mem.allocated().len() == 10);
-}
-
-fn grow_zeroed<T>(mut mem: impl RawMem<Item = T>) {
-    unsafe {
-        mem.grow_zeroed(10).expect("grow");
-    }
-    assert!(mem.allocated().len() == 10);
-}
-
-fn grow_assumed<T>(mut mem: impl RawMem<Item = T>) {
-    unsafe {
-        mem.grow_assumed(10).expect("grow");
-    }
-    assert!(mem.allocated().len() == 10);
 }
